requirements.txt
requests>=2.0.0
psycopg2-binary==2.9.6
